requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
cx-Oracle==7.2.2
et-xmlfile==1.0.1
google-api-core==1.14.2
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.11.2
google-cloud-core==0.29.1
google-cloud-datastore==1.7.3
google-cloud-logging==1.10.0
google-cloud-pubsub==0.45.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
grpc-google-iam-v1==0.12.3
grpcio==1.23.0
httplib2==0.13.1
idna==2.8
jdcal==1.4.1
numpy==1.16.3
openpyxl==2.6.3
pandas==0.24.2
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pyodbc==4.0.27
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.6
uritemplate==3.0.0
urllib3==1.25.3
wincertstore==0.2
